Add result operation with TDD and small refactor
